S-24: Moved date format functions to date format service, refactored affected code.
