+++ closetr-web/src/app/pipes/date-range-filter.pipe.ts
@@ -1,9 +1,15 @@
 import { Pipe, PipeTransform } from '@angular/core';
+import { DateFormatService } from '../services/utils/date-format.service';
 
 @Pipe({
   name: 'dateRangeFilter'
 })
 export class DateRangeFilterPipe implements PipeTransform {
+  dateFormatService: DateFormatService;
+
+  constructor(private dateformatservice: DateFormatService) {
+    this.dateFormatService = dateformatservice;
+  }
 
   transform(items: any, dateFrom: Date, dateTo: Date, property: string): any {
     if (!dateFrom || !dateTo) {
@@ -14,7 +20,7 @@ export class DateRangeFilterPipe implements PipeTransform {
     let filteredItems = {};
     for (let itemID in items) {
       let item = items[itemID];
-      let itemDateProperty = this.formatStringDate(item[property]);
+      let itemDateProperty = this.dateFormatService.formatStringDate(item[property]);
       //comparing item date and criteria
       if (dateFrom <= itemDateProperty && dateTo >= itemDateProperty) {
         filteredItems[itemCount] = item;
@@ -22,16 +28,6 @@ export class DateRangeFilterPipe implements PipeTransform {
       }
     }
     return filteredItems;
-  }
-
-  /*
-  format string to date, from ISO format.
-  */
-  formatStringDate(date: string): Date {
-    var year = parseInt(date.substring(0,4));
-    var month = parseInt(date.substring(5,7)) - 1;
-    var day = parseInt(date.substring(8,10));
-    return (new Date(year, month, day));
   }
 
 }
